fixed error processing

+++ app/schemas.py
@@ -17,7 +17,7 @@ class MovieSchema(Schema):
     description = fields.Str()
     trailer = fields.Str()
     year = fields.Int()
-    rating = fields.Int()
+    rating = fields.Float()
     genre_id = fields.Int()
     genre = fields.Nested(GenreSchema)
     director_id = fields.Int()
